Move AchievementService to utils-backend so it can be imported
