+++ packages/engine/paima-rest/package.json
@@ -7,8 +7,8 @@
   "types": "build/index.d.ts",
   "scripts": {
     "lint:eslint": "eslint .",
-    "build": "tsc --build tsconfig.build.json",
-    "prebuild": "npm run compile:api",
+    "build": "npm run compile:api && tsc --build tsconfig.build.json",
+    "prebuild": "",
     "compile:api": "tsoa spec-and-routes"
   },
   "author": "Paima Studios",
