+++ packages/engine/paima-rest/src/EngineService.ts
@@ -1,9 +1,12 @@
 import type { GameStateMachine } from '@paima/sm';
+import { AchievementService } from '@paima/utils-backend';
 
 export class EngineService {
   public static INSTANCE = new EngineService();
 
   private runtime: GameStateMachine | undefined = undefined;
+
+  public achievementService: AchievementService = new AchievementService();
 
   getSM = (): GameStateMachine => {
     if (this.runtime == null) {
