+++ packages/engine/paima-rest/tsoa.json
@@ -13,7 +13,8 @@
   "compilerOptions": {
     "baseUrl": "./",
     "paths": {
-      "@paima/utils/*": ["../../paima-sdk/paima-utils/*"]
+      "@paima/utils": ["../../paima-sdk/paima-utils/build"],
+      "@paima/node-sdk/paima-utils-backend": ["../../node-sdk/paima-utils-backend/build"]
     }
   }
 }
